Multipart upload works
Reading data in 100M chunks ensures that only 100M are in memory and in
transit at a time. I thought about uploading in parallel but that would
require threads within threads, which would be more complex to
understand. Instead, I realized that files are getting uploaded in a
serialized form anyway, so no need to add complexity