Improve speed of tablechop
This paralellises tablechop's fetching of the .json files to reduce the
amount of time it takes to run. It adds a dependency on eventlet.

This also removes the `keep_existing` option in favour of a different
approach - it wasn't really clear to me what the intention of
`keep_existing` was:

Instead of keeping all files which exist on disk, plus all json files
which reference it, this update will always keep the last backup.