Listen for IN_CLOSE_WRITE event by default
Closes #42.

Cassandra 2.x adds a file per sstable ending in `-Summary.db`. This is
written directly in to the data directory, so doesn't trigger the
IN_MOVED_TO event.

These files can be regenerated by Cassandra but it speeds up recovery if
they're backed up. More generally, I don't really see a reason why we
shouldn't be listening for this event, as we should probably be
backing up everything by default in future.